+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 0.0.2
+Feature:
+
+* send event `dataConnectionClose` when data connection was close
+* methode for get all open data connections
+
 ## 0.0.1
 first version